Updated build script

+++ install.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 npm install
 bower install
-cp config/main.sample.js config/main.js
+if [ ! -f config/main.js ]; then
+    cp config/main.sample.js config/main.js
+fi
